requirements.txt
Django==3.2.16
mixer==7.2.2
pytest==7.1.3
pytest-django==4.5.2
flake8==5.0.4
django-debug-toolbar==3.8.1